refactor: move DataManager in a standalone package
