+++ app/src/main/java/com/osfans/trime/data/sound/SoundEffectManager.kt
@@ -8,7 +8,7 @@ import androidx.annotation.Keep
 import com.charleskorn.kaml.Yaml
 import com.charleskorn.kaml.YamlConfiguration
 import com.osfans.trime.data.AppPrefs
-import com.osfans.trime.data.DataManager
+import com.osfans.trime.data.base.DataManager
 import timber.log.Timber
 import java.io.File
