+++ app/src/main/java/com/osfans/trime/data/theme/ColorManager.kt
@@ -11,7 +11,7 @@ import android.graphics.drawable.GradientDrawable
 import androidx.annotation.ColorInt
 import androidx.core.math.MathUtils
 import com.osfans.trime.data.AppPrefs
-import com.osfans.trime.data.DataManager
+import com.osfans.trime.data.base.DataManager
 import com.osfans.trime.data.sound.SoundEffectManager
 import com.osfans.trime.util.ColorUtils
 import com.osfans.trime.util.WeakHashSet
