+++ app/src/main/java/com/osfans/trime/data/theme/FontManager.kt
@@ -9,7 +9,7 @@ import android.graphics.fonts.Font
 import android.graphics.fonts.FontFamily
 import android.os.Build
 import androidx.annotation.RequiresApi
-import com.osfans.trime.data.DataManager
+import com.osfans.trime.data.base.DataManager
 import timber.log.Timber
 import java.io.File
